settings_modal
settings_modal changes.  Is this supposed to be a modal or a page?  It
is currently being called in the navigation as a page.  That would need
to change if it is supposed to be a modal.  Whatever the case I added
the styling like the modals and pulled in that modal scss.  Got rid of
the p class="welcome" and went with a h4 instead.